Changes the Coumn to Row
Changes the Coumn Widget to Row Row Widget and the things flip over. The Row takes full Width but not the full height, in order to do that you have to specify the height in the SizedBox Widget. In this case for full screen it had to be double.infinity.